Make sure PDLA::Lite always exports core functions.
Otherwise only the first caller would have these functions loaded, and it would thus not match the documentation.

This issue has been around for a long time, as it pre-dates the addition of the package statement in 87487f0c4abd51268b898752d9b3029b926ec7aa .

Also update the documentation to be clear about which methods are exported, and encourage the method calling syntax regardless.

An alternative is to export nothing and document the method syntax as the only approach, but that would lead to CPAN breakage for those that rely on pdl() etc being exported (albeit such packages will not work if they are loaded second anyway).

Updates issue #11