Make cat return a piddle of the highest input type, not the first.
Previously the first element of a list provided to 'cat' was used
to set the type of the output piddle.  Now it uses the "highest"
type if passed a mixed-type list.

Also added some tests for this new behavior, and improved the
documentation for cat, append, and glue to better highlight the
differences between then.